Create The painter
We have to paint n boards of length {A1, A2…An}. There are k painters available and each takes 1 unit of time to paint 1 unit of the board. The problem is to find the minimum time to get this job was done under the constraints that any painter will only paint continuous sections of boards, say board {2, 3, 4} or only board {1} or nothing but not board {2, 4, 5}.

NOTE :-

After answering the question, attempt the related question in the linked resource to improve your understanding of the question . Click here

Input Format

An integer N as size of array.

N integer represents elements of the array.

An integer k .

Constraints

1 <= N <= 10^4
1 <= arr[i] <= 10^4
1 <=  k <= 10^4
Output Format

print the final output.

Sample Input 0

4
10 10 10 10 
2    
Sample Output 0

20
Explanation 0

Here we can divide the boards into 2 equal sized partitions, so each painter gets 20 units of board and the total time taken is 20.