Create Weighty Voyage
In order to transport packages from one port to another within a specific time frame, there is a conveyor belt that contains these packages. Each package on the conveyor belt is assigned a weight denoted by weights[i]. The packages are loaded onto the ship on a daily basis in the order given by the weights, ensuring that the total weight loaded onto the ship does not exceed the maximum weight capacity of the ship.

Your task is to determine the minimum weight capacity that the ship needs to have in order to successfully transport all the packages on the conveyor belt within the given time frame of D days.

Input Format

first line contains integer N representing the size of array.

second line contains N integers representing th elements of array.

third line contains integer D representing days.

Constraints

1 <= days <= N <= 5 * 10^4 1 <= arr[i] <= 500

Output Format

Return the least weight capacity of the ship that will result in all the packages on the conveyor belt being shipped within D days.

Sample Input 0

10
1 2 3 4 5 6 7 8 9 10
5
Sample Output 0

15
Explanation 0

A ship capacity of 15 is the minimum to ship all the packages in 5 days like this: 1st day: 1, 2, 3, 4, 5 2nd day: 6, 7 3rd day: 8 4th day: 9 5th day: 10

Note that the cargo must be shipped in the order given, so using a ship of capacity 14 and splitting the packages into parts like (2, 3, 4, 5), (1, 6, 7), (8), (9), (10) is not allowed.