Create Minimum Add to Make Parentheses Valid
A parentheses string is valid if and only if:

It is the empty string, It can be written as AB (A concatenated with B), where A and B are valid strings, or It can be written as (A), where A is a valid string. You are given a parentheses string s. In one move, you can insert a parenthesis at any position of the string.

For example, if s = "()))", you can insert an opening parenthesis to be "(()))" or a closing parenthesis to be "())))". Return the minimum number of moves required to make s valid.

Input Format

First line contains a string s.

Constraints

1 <= s.length <= 1000 s[i] is either '(' or ')'.

Output Format

Returns an integer value.

Sample Input 0

())
Sample Output 0

1
Explanation 0

Only 1 opening bracket will make the string s valid "(())"