Create Facing the sun
Given an array H representing heights of buildings. You have to count the buildings which will see the sunrise (Assume : Sun rise on the side of array starting point).image

Input Format

array of integers h and n as the size of array.

Constraints

1 ≤ N ≤ 106 1 ≤ Hi ≤ 108 Expected Time Complexity: O(N) Expected Auxiliary Space: O(1)

Output Format

integer representing the number of buildings which will see the sunrise

Sample Input 0

5
7 4 8 2 9
Sample Output 0

3
Explanation 0

As 7 is the first element, it can see the sunrise. 4 can't see the sunrise as 7 is hiding it.
8 can see as it's greater than 7. 2 can't see the sunrise. 9 can also see the sunrise.

Sample Input 1

4
2 3 4 5
Sample Output 1

4
Explanation 1

As 2 is the first element, it can see the sunrise. 3 can see the sunrise as 2 is not hiding it. Same for 4 and 5, they also can see the sunrise.