Create weakest rows
You are given an m x n binary matrix mat of  (representing soldiers) and  (representing civilians). The soldiers are positioned in front of the civilians. That is, all the  will appear to the left of all the  in each row.

A row  is weaker than a row  if one of the following is true:

The number of soldiers in row  is less than the number of soldiers in row .
Both rows have the same number of soldiers and .
Return the indices of the  weakest rows in the matrix ordered from weakest to strongest.

NOTE :- After answering the question, attempt the related question in the linked resource to improve your understanding of this question . Click here

Input Format

Three Integers , , .
 lines of  Integers spaces with single space.
Constraints

m == mat.length
n == mat[i].length
2 <= n, m <= 10000
1 <= k <= m
matrix[i][j] is either 0 or 1.
Output Format

 Integers spaced with single space.

Sample Input 0

5 5 3
1 1 0 0 0
1 1 1 1 0
1 0 0 0 0 
1 1 0 0 0 
1 1 1 1 1
Sample Output 0

2 0 3
Explanation 0

The number of soldiers in each row is:

Row 0: 2
Row 1: 4
Row 2: 1
Row 3: 2
Row 4: 5
The rows ordered from weakest to strongest are [2,0,3,1,4].